Update Alpine to latest 3.12.3 release
This upgrades Alpine from 3.11.x releases (which will end support
2021-11-01) to 3.12.x (which will be supported until 2022-05-01)

The most current Alpine release is 3.13.x.  Unfortunately, that release
brings in php 7.4.x, which currently does not work with all Islandora
PHP dependencies.

For some reason, the public keys generated when building ImageMagick do
not work as expected, so `apk` will complain about untrusted signatutres
and fail to install the packages that are custom-built in the
ImageMagick images.  As a workaround, --allow-untrusted flag is used
when installing the custom-built ImageMagick.  We trust ourselves, so
this is likely OK.