Update starter site & matomo extra tools & fix permission Bugs (#350)
* Update Matomo Extra Tools from 5.0.1 -> 5.0.4

Required to fix build issue, that was being
covered up by a previously successful cache.

* Fix for potential permission issues.

A number of permission issues have gone unnoticed,
I became aware of them when I did a fresh build
without cache on a newly installed computer.

When building Docker images, Docker copies
files/folders with the same permissions as they
have on the host machine. Git does not track the
permissions of files except for the executable bit
on files. So depending on who builds the images
files/folders can get different permissions.

This changes it such that the Makefile will ensure
the appropriate permissions are set when building.

This is required as we run services as different
users other than root in the container. Those
users must be able to access and execute certain
files.

* Update to latest starter site, and hard-code paths to support it.